Utils: Move battle-related methods to a dedicated class
