+++ src/lib/Utils/Gym.js
@@ -35,7 +35,7 @@ class AutomationUtilsGym
         let bestGymRate = 0;
 
         let playerClickAttack = App.game.party.calculateClickAttack();
-        let totalAtkPerSecondByRegion = Automation.Utils.Route.getPlayerWorstAttackPerSecondForAllRegions(playerClickAttack);
+        let totalAtkPerSecondByRegion = Automation.Utils.Battle.getPlayerWorstAttackPerSecondForAllRegions(playerClickAttack);
 
         for (const gymData of this.__internal__gymGemTypeMap.get(pokemonType))
         {
@@ -57,7 +57,7 @@ class AutomationUtilsGym
                     continue;
                 }
 
-                let currentPokemonTickToDefeat = Automation.Utils.Route.getGameTickCountNeededToDefeatPokemon(
+                let currentPokemonTickToDefeat = Automation.Utils.Battle.getGameTickCountNeededToDefeatPokemon(
                     pokemon.maxHealth, playerClickAttack, totalAtkPerSecondByRegion[gymData.region]);
                 currentGymGemPerTick += (GameConstants.GYM_GEMS / currentPokemonTickToDefeat);
             }
