+++ tst/imports/AutomationUtils.import.js
@@ -1,4 +1,5 @@
 // Import current lib elements
+import "src/lib/Utils/Battle.js";
 import "src/lib/Utils/LocalStorage.js";
 import "src/lib/Utils/Gym.js";
 import "src/lib/Utils/OakItem.js";
